+++ megan_global_explanations/utils.py
@@ -215,4 +215,24 @@ class RecordIntermediateEmbeddingsCallback(EpochCounterCallback):
             self.logger.info(f' * stored intermediate embeddings at epoch {self.epoch}')
             
             
-            
+# == GRAPH UTILITY ==
+
+def extend_graph_info(index_data_map: dict):
+    """
+    This function extends the information contained in the graph dicts within the given ``index_data_map`` 
+    representation of a visual graph dataset. In reality, this function will simply copy certain information
+    from the metadata of the elements direclty into the graph dict, which is needed for some processing 
+    steps.
+    
+    :param index_data_map: The visual graph dataset representation
+    
+    :returns: None, the modification is done in place.
+    """
+    for index, data in index_data_map.items():
+        graph = data['metadata']['graph']
+        graph['graph_index'] = data['metadata']['index']
+        
+        if 'repr' in data['metadata']:
+            graph['graph_repr'] = data['metadata']['repr']
+        else:
+            graph['graph_repr'] = data['metadata']['value']